break out CustomPropertyInspector class to its own file
